Character Selection Scene Update
Character Selection Scene works with no bugs for now
CharacterSelection script and SelectionManager are both used to update
the displayed picture, cached player properties, and displayed names.
NOTE: Do not try to click the BACK button when you're player 2 and in
the StageSelection scene, it will cause an error as of now.